% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon2amino.R
\name{codon2amino}
\alias{codon2amino}
\title{Title: Using the codon table identify the codons and their respective amino acid, return the aminoacids}
\usage{
codon2amino(codons)
}
\arguments{
\item{codons}{codons from F3}
}
\value{
amino_seq 1 letter code AA
}
\description{
Title: Using the codon table identify the codons and their respective amino acid, return the aminoacids
}
