% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Convert DNA to RNA}
\usage{
dna_to_rna(dna_seq)
}
\arguments{
\item{dna_seq}{The DNA sequence that is going to be converted into RNA.}
}
\value{
Sequence provided converted into RNA.
}
\description{
Convert a DNA sequence into a RNA sequence by replacing the Thymines (T) by
Uracil (U) present in the given sequence.
}
\examples{
dna_seq = "AGAGTTTGATCCTGGCTCAG"
dna_to_rna(dna_seq)
}
