% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_count.R
\name{seq_count}
\alias{seq_count}
\title{seq_count}
\usage{
seq_count(seq)
}
\arguments{
\item{seq}{character vector containing amino acid sequence.}
}
\value{
A bar plot showing how many times each amino acid occurs in
the sequence.
}
\description{
Count occurrences of each amino acid in sequence
}
\examples{
aa_seq <- "GWIFGTTLDSKTQSLLIVNNATNVVIKVCEFQFCNDPFLGVYYHKNNKSWMESEFRVY"
seq_count(aa_seq)

}
